+++ CHANGELOG.md
@@ -6,8 +6,10 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+- Enabled support for Nylas API v2.2
 - Add `metadata` field in the Event model to support new event metadata
 - Add support for filtering `metadata` using `metadata_key`, `metadata_value`, and `metadata_pair`]
+- Add support for new `RoomResource` fields
 - Add missing getters for Event.Recurrence fields
 - Add Neural API support